+++ packages/adhere-ui-anthoc/lib/calendar-timestamp-value-hoc/CalendarTimestampValueHOC.d.ts
@@ -1,0 +1,6 @@
+import type { CalendarFormatValueHOCProps } from '../types';
+import ValueHOC from './ValueHOC';
+declare const CalendarFormatValueHOC: typeof ValueHOC & {
+    defaultProps?: Partial<CalendarFormatValueHOCProps>;
+};
+export default CalendarFormatValueHOC;