+++ packages/adhere-ui-anthoc/src/calendar-timestamp-value-hoc/index.ts
@@ -1,0 +1,3 @@
+import CalendarTimestampValueHOC from './CalendarTimestampValueHOC';
+
+export default CalendarTimestampValueHOC;